+++ .gitignore
@@ -43,7 +43,7 @@ npm-debug.log
 
 # Stored hash of source tracker files used in development environment
 # to detect changes in /tracker/src and avoid unnecessary compilation.
-/tracker/dev-compile/last-hash.txt
+/tracker/compiler/last-hash.txt
 
 # test coverage directory
 /assets/coverage
