+++ tracker/compiler/can-skip-compile.js
@@ -4,7 +4,7 @@ const crypto = require('crypto');
 
 const LAST_HASH_FILEPATH = path.join(__dirname, './last-hash.txt')
 
-// Re-compilation is only required if any of these files have been changed. 
+// Re-compilation is only required if any of these files have been changed.
 const COMPILE_DEPENDENCIES = [
   path.join(__dirname, '../compile.js'),
   path.join(__dirname, '../src/plausible.js'),
